AG-9040 - Update range bar series docs
