+++ packages/ag-charts-website/src/content/docs/range-bar-series/_examples/horizontal-range-bar/main.ts
