+++ packages/ag-charts-website/src/content/docs/range-bar-series/_examples/multiple-horizontal-range-bars/data.ts
@@ -1,0 +1,28 @@
+export function getData() {
+    return [
+        {
+            smartphone: 'Apple iPhone',
+            lowCost: 188,
+            highCost: 501,
+            lowRetail: 599,
+            highRetail: 1449,
+            profitMargin: 68.6
+        },
+        {
+            smartphone: 'Google Pixel',
+            lowCost: 285.75,
+            highCost: 485.50,
+            lowRetail: 699,
+            highRetail: 899,
+            profitMargin: 54.6
+        },
+        {
+            smartphone: 'Samsung Galaxy',
+            lowCost: 213,
+            highCost: 670,
+            lowRetail: 549,
+            highRetail: 1799,
+            profitMargin: 59.2
+        },
+    ];
+}
