+++ packages/ag-charts-website/src/content/docs/range-bar-series/_examples/multiple-range-bars/main.ts
@@ -1,106 +1,81 @@
-import { AgChartOptions, AgEnterpriseCharts } from 'ag-charts-enterprise';
+import { AgChartOptions, AgEnterpriseCharts, AgRangeBarSeriesLabelFormatterParams } from 'ag-charts-enterprise';
 import { getData } from './data';
 
 const options: AgChartOptions = {
     container: document.getElementById('myChart'),
     data: getData(),
     title: {
-        text: 'Smartphone Production Cost Vs Retail Price',
+        text: `Pete's Instagram Followers`,
     },
     subtitle: {
-        text: 'Production cost range vs retail price range of top-selling phone brands on the market in 2023',
-        spacing: 30
+        text: `He Wishes`
     },
-        footnote: {
-        text: 'Costs include essential components like core processors, display, memory, and camera module but exclude marketing, research, distribution, staff, accessories, packaging, and software.',
-        spacing: 30
-    },
-    series: [
+    series:[
         {
             type: 'range-bar',
-            direction: 'horizontal',
-            xKey: 'smartphone',
-            xName: 'Smartphone',
-            yLowKey: 'lowCost',
-            yHighKey: 'highCost',
-            yLowName: 'Lowest Cost',
-            yHighName: 'Highest Cost',
-            yName: 'Production Cost Range',
-            fill: '#D1C0A8',
-            stroke: '#D1C0A8',
-        },
-        {
-            type: 'range-bar',
-            direction: 'horizontal',
-            xKey: 'smartphone',
-            xName: 'Smartphone',
-            yLowKey: 'lowRetail',
-            yHighKey: 'highRetail',
-            yLowName: 'Lowest Price',
-            yHighName: 'Highest Price',
-            yName: 'Retail Price Range',
+            xKey: 'date',
+            xName: 'Date',
+            yLowKey: 'start',
+            yHighKey: 'gain',
+            yLowName: 'Start',
+            yHighName: 'End',
+            yName: 'Followers Gained',
             fill: '#205C37',
             stroke: '#205C37',
+            label: {
+                placement: 'outside',
+                formatter: ({ itemId, yLowValue, yHighValue }: AgRangeBarSeriesLabelFormatterParams) => {
+                    return itemId === 'low'
+                        ? ''
+                        : `+${(yHighValue - yLowValue).toFixed(0)}`;
+                },
+            },
         },
         {
-            type: 'scatter',
-            yKey: 'smartphone',
-            xKey: 'profitMargin',
-            xName: 'Profil Margin',
-            yName: 'Profit Margin %',
-            sizeKey: 'profitMargin',
-            labelKey: 'profitMargin',
+            type: 'range-bar',
+            xKey: 'date',
+            xName: 'Date',
+            yLowKey: 'loss',
+            yHighKey: 'gain',
+            yLowName: 'End',
+            yHighName: 'Start',
+            yName: 'Followers Lost',
+            fill: '#D1C0A8',
+            stroke: '#D1C0A8',
             label: {
-                formatter: ({value}) => `${(+value).toFixed(0)}%`
+                placement: 'outside',
+                formatter: ({ itemId, yLowValue, yHighValue }: AgRangeBarSeriesLabelFormatterParams) => {
+                    return itemId === 'high'
+                        ? ''
+                        : `-${(yHighValue - yLowValue).toFixed(0)}`;
+                },
             },
+        },
+        {
+            type: 'line',
+            xKey: 'date',
+            xName: 'Date',
+            yKey: 'loss',
+            yName: 'Net Followers',
+            stroke: '#5c2045',
+            strokeWidth: 4,
             marker: {
-                // fill: '#5c2045',
-                // stroke: '#5c2045',
-                stroke: '#205C37',
-                fill: '#ced1a8',
-                strokeWidth: 1,
-                maxSize: 50,
-                size: 15
+                fill: '#5c2045',
+                stroke: '#5c2045',
+                size: 0,
             },
         },
     ],
     axes: [
         {
             type: 'category',
-            position: 'left',
-            keys: ['smartphone'],
-            groupPaddingInner: 0,
-            paddingInner: 0.9,
-            paddingOuter: 0.8,
-        },
-        {
-            type: 'number',
-            position: 'top',
-            keys: ['profitMargin'],
-            label: {
-                formatter: ({ value }) => `${value}%`,
-            },
-            line: {
-                width: 0,
-            },
-        },
-        {
-            type: 'number',
             position: 'bottom',
-            keys: ['lowRetail', 'highRetail', 'lowCost', 'highCost'],
-            label: {
-                formatter: ({ value }) =>
-                    `${(+value).toLocaleString('en-US', {
-                        style: 'currency',
-                        currency: 'USD',
-                        maximumFractionDigits: 0,
-                    })}`,
-            },
-            line: {
-                width: 0,
-            },
-        },
-    ],
+            groupPaddingInner: 0
+        }, {
+            type: 'number',
+            position: 'left'
+        }
+    ]
 };
 
 AgEnterpriseCharts.create(options);
