--- packages/ag-charts-website/src/content/docs/range-bar-series/_examples/multiple-range-columns/data.ts
@@ -1,76 +1,0 @@
-export function getData() {
-    return [
-        {
-            date: 'Jan',
-            start: 0,
-            gain: 210,
-            loss: 200,
-        },
-        {
-            date: 'Feb',
-            start: 200,
-            gain: 440,
-            loss: 305,
-        },
-        {
-            date: 'Mar',
-            start: 305,
-            gain: 589,
-            loss: 488,
-        },
-        {
-            date: 'Apr',
-            start: 488,
-            gain: 809,
-            loss: 721,
-        },
-        {
-            date: 'May',
-            start: 721,
-            gain: 1002,
-            loss: 833,
-        },
-        {
-            date: 'Jun',
-            start: 833,
-            gain: 1035,
-            loss: 852,
-        },
-        {
-            date: 'Jul',
-            start: 852,
-            gain: 1050,
-            loss: 852,
-        },
-        {
-            date: 'Aug',
-            start: 852,
-            gain: 1059,
-            loss: 924,
-        },
-        {
-            date: 'Sep',
-            start: 924,
-            gain: 1205,
-            loss: 1115,
-        },
-        {
-            date: 'Oct',
-            start: 1115,
-            gain: 1390,
-            loss: 1231,
-        },
-        {
-            date: 'Nov',
-            start: 1231,
-            gain: 1425,
-            loss: 1306,
-        },
-        {
-            date: 'Dec',
-            start: 1306,
-            gain: 1498,
-            loss: 1484,
-        },
-    ];
-}
