+++ packages/ag-charts-website/src/content/docs/range-bar-series/_examples/simple-range-bar/data.ts
@@ -1,4 +1,3 @@
-// Source: http://www.sheffieldweather.co.uk/Averages/MONTHLYAIRAVERAGE.htm
 export function getData() {
     return [
         {
