+++ packages/ag-charts-website/src/content/docs/range-bar-series/_examples/simple-range-bar/main.ts
